+++ src/chart/rendering/drawinterlacing.cpp
@@ -25,8 +25,9 @@ void drawInterlacing::draw(bool horizontal, bool text)
 {
 	auto axisIndex = horizontal ? Diag::ScaleId::y : Diag::ScaleId::x;
 
-	if ((*style.plot.getAxis(axisIndex).interlacing.color).alpha <= 0.0)
-		return;
+	auto interlacingColor = *style.plot.getAxis(axisIndex).interlacing.color;
+
+	if (!text && interlacingColor.alpha <= 0.0) return;
 
 	const auto &axis = diagram.axises.at(axisIndex);